Extracted TeamSlackChannelService
